requirements.txt
# Minimal dependencies for Kindle Newsletter Prototype
beautifulsoup4==4.12.2
ebooklib==0.18
Pillow==10.1.0
readability-lxml==0.8.1
requests==2.31.0
flask==2.3.3
gunicorn==21.2.0
sendgrid==6.11.0
python-dotenv==1.0.0
